requirements.txt
streamlit==1.28.1
torch==2.6.0
torchvision==0.21.0
numpy==2.0.2
Pillow==10.0.1
matplotlib==3.10.0
